+++ src-test/saltchannel/v2/packets/M2PacketTest.java
@@ -7,14 +7,14 @@ public class M2PacketTest {
 
     @Test
     public void testSanity() {
-        M2Packet p = new M2Packet();
+        M2Message p = new M2Message();
         p.noSuchServer = false;
         p.serverEncKey = new byte[32];
 
         byte[] bytes1 = new byte[p.getSize()];
         p.toBytes(bytes1, 0);
         
-        M2Packet p2 = M2Packet.fromBytes(bytes1, 0);
+        M2Message p2 = M2Message.fromBytes(bytes1, 0);
         
         byte[] bytes2 = new byte[bytes1.length];
         p2.toBytes(bytes2, 0);
@@ -24,7 +24,7 @@ public class M2PacketTest {
     
     @Test
     public void testResumeSupported() {
-        M2Packet p = new M2Packet();
+        M2Message p = new M2Message();
         p.noSuchServer = false;
         p.resumeSupported = true;
         p.serverEncKey = new byte[32];
@@ -32,7 +32,7 @@ public class M2PacketTest {
         byte[] bytes1 = new byte[p.getSize()];
         p.toBytes(bytes1, 0);
         
-        M2Packet p2 = M2Packet.fromBytes(bytes1, 0);
+        M2Message p2 = M2Message.fromBytes(bytes1, 0);
         
         byte[] bytes2 = new byte[bytes1.length];
         p2.toBytes(bytes2, 0);