+++ release-notes.md
@@ -5,13 +5,25 @@ Release history. The implementation started in 2016, but formal
 releases starts with release 2.0.
 
 
-Release v2.1, 2017-11-17
+
+Release v2.2, 2017-10-25
+========================
+
+Implements salt-channel-v2-draft8.md. No known bugs.
+
+* ApplicationChannel now available to lib users. Includes
+  information about last flag and multi app packet.
+
+
+
+Release v2.1, 2017-10-17
 ========================
 
 This release implements Salt Channel v2, draft8.
 
 
-Release v2.0, 2017-11-17
+
+Release v2.0, 2017-10-17
 ========================
 
 Implements Salt Channel v2, draft7 (spec-salt-channel-v2-draft7.md).